Reorganize directory hierarchy
Be more explicit about the purpose of each config file. While the repo
is still young, shuffling config files around is still okay to do.

Given that I don't have any rules for the `protecting` category, add a
.gitignore that ignores everything in that directory (excluding itself).
On the first commit adding a rule for the `protecting` category, remove
the .gitignore file.

Provide better documentation surrounding the project in README.md.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>